+++ docs/nuxt.config.ts
@@ -7,12 +7,9 @@ export default withDocus({
   },
   loading: { color: '#ff4785' },
   buildModules: [
-    // https://github.com/bdrtsky/nuxt-ackee
-    'nuxt-ackee'
+    'vue-plausible'
   ],
-  ackee: {
-    server: 'https://ackee.nuxtjs.com',
-    domainId: '7e54dba4-1702-4c2a-9fb4-e0657449fb95',
-    detailed: true
+  plausible: {
+    domain: 'storybook.nuxtjs.org'
   }
 })