+++ examples/manual-setup/.storybook/middleware.js
@@ -1,0 +1,1 @@
+module.exports = require('../.nuxt-storybook/storybook/middleware.js')