+++ examples/tailwind/components/MyButton.stories.js
@@ -1,5 +1,26 @@
-export default { title: 'Button' }
 
-export const primaryButton = () => '<MyButton>Primary Button</MyButton>'
-export const secondaryButton = () => '<MyButton type="secondary">Secondary Button</MyButton>'
-export const roundButton = () => '<MyButton round>Round Button</MyButton>'
+import MyButton from './MyButton'
+export default {
+    title: 'Button',
+    component: MyButton,
+    argTypes: {
+        type: {
+          control: {
+            type: 'select',
+            options: ['primary', 'secondary'],
+          },
+          defaultValue: 'primary'
+        },
+        round: { control: 'boolean' },
+        text: {
+            control: 'text',
+            defaultValue: 'Awesome Button'
+        }
+    }
+}
+
+export const Button = (arg, { argTypes }) => ({
+    components: { MyButton },
+    props: Object.keys(argTypes),
+    template: '<MyButton v-bind="$props" />'
+})