+++ src/runtime/middlewares.js
@@ -1,0 +1,8 @@
+const middlewares = []
+
+module.exports = {
+  middlewares,
+  addServerMiddleware: (middleware) => {
+    middlewares.push(middleware)
+  }
+}