+++ storybook/main.js
@@ -1,5 +1,13 @@
 const stories = [<%= options.stories.map(s => `'${s}'`).join(",") %>]
-const addons  = [<%= options.addons.map(s => devalue(s)).join(",") %>]
+const addons  = [<%= options.addons.map(s => devalue(s)).join(",") %>,{
+  name: '@storybook/addon-postcss',
+  options: {
+    postcssLoaderOptions: {
+      implementation: require('postcss')
+    }
+  }
+}
+]
 
 function nuxifyStorybook(storybookConfig) {
   return {